extract jwt to jwt.ts
