+++ tsconfig.json
@@ -14,6 +14,6 @@
       "dom"
     ]
   },
-  "include": ["src"],
+  "include": ["src/*.ts"],
   "exclude": ["node_modules", "**/__tests__/*"]
 }
